[#70] refactor: 블로그 크롤링 로직 리팩토링
